schedule.c를 create.c와 delete.c로 분리
